refactor: hooks関数をファイルにまとめた
