+++ frontend/src/hooks/api/relations/useBlockUsers.ts
@@ -6,3 +6,13 @@ export const useBlockUsers = (): { users: User[] } => {
 
   return { users };
 };
+
+export const useIsBlockedUser = (
+  userId: string
+): { isBlockedUser: boolean } => {
+  const { users: blockedUsers } = useBlockUsers();
+  const isBlockedUser =
+    blockedUsers.find((blockedUser) => blockedUser.id === userId) !== undefined;
+
+  return { isBlockedUser };
+};
